Omaha #3410 changed lon/lat fields to floats in sfc and air obs location

Former-commit-id: 12ec17f15843710be75a12658ebb3208fa9b4b1e [formerly 12ec17f15843710be75a12658ebb3208fa9b4b1e [formerly a7d17bd490f2f0931f3c22ee5484214041a34bd6]] [formerly aa991c922354f0a67fac08fee456ca60a30de968] [formerly 8dfe4ed6f2fc5224fb344df59f22ff0dd1ea6277 [formerly 8dfe4ed6f2fc5224fb344df59f22ff0dd1ea6277 [formerly 74a6c3dd20898e9bd4484376ea4e4e471bdec9af]] [formerly aa991c922354f0a67fac08fee456ca60a30de968 [formerly aa991c922354f0a67fac08fee456ca60a30de968 [formerly cdb0a72fcb2345564f2b944fdecd7869fd498767]]]]
Former-commit-id: 8dfe4ed6f2fc5224fb344df59f22ff0dd1ea6277 [formerly 8dfe4ed6f2fc5224fb344df59f22ff0dd1ea6277 [formerly 74a6c3dd20898e9bd4484376ea4e4e471bdec9af]]
Former-commit-id: 3b341a078dee5ac0d23ca2cf4fcb5758127f0d34 [formerly 8623a5a72eaa9e27e6806554ca01fa36d0dc76b3] [formerly 995bdcaabc791724badcf4808e4b2ad68b74e80d [formerly 10ea02b9366284a7b97f57ffdb7ac54688cfa0c4]]
Former-commit-id: 507b40cc5ea8262df5139778a477be34fd3a4a6d [formerly 5977b049c1eda298ce206a8d5fc34b8f235639c5]
Former-commit-id: 7b89dac8ace4878dbe1d2f95cd802b1268b7bb30

+++ cave/com.raytheon.uf.viz.npp.nucaps/src/com/raytheon/uf/viz/npp/nucaps/provider/NucapsSoundingProvider.java
@@ -65,6 +65,7 @@ import com.vividsolutions.jts.index.strtree.STRtree;
  * Jul 25, 2013 2190       mschenke    Initial creation
  * Aug 15, 2013 2260       bsteffen    Switch poessounding to NSharp.
  * Aug 27, 2013 2190       mschenke    Fixed point query request
+ * Jul 23, 2014 3410       bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -114,8 +115,8 @@ public class NucapsSoundingProvider extends
                 sounding.setStationId(NucapsRecord.PLUGIN_NAME);
                 Coordinate location = getSoundingLocation();
                 sounding.setName(GeoUtil.formatCoordinate(location));
-                sounding.setLongitude(location.x);
-                sounding.setLongitude(location.y);
+                sounding.setLongitude((float) location.x);
+                sounding.setLongitude((float) location.y);
 
                 // Get pressure values
                 Number[] pressures = pdv
